Split some of the large files into smaller files and start to organise into sub-directories
